+++ plugin.php
@@ -3,7 +3,7 @@
  * MaterialsPlugin Name: TMS Materials
  * MaterialsPlugin URI: https://github.com/devgeniem/tms-plugin-materials
  * Description: TMS Materials
- * Version: 0.0.1
+ * Version: 1.0.0
  * Requires PHP: 7.4
  * Author: Geniem Oy
  * Author URI: https://geniem.com